Added rowmapper as a bean.
